+++ samples/hello-world/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion 30
 
     defaultConfig {
-        applicationId "com.samples.helloworld"
+        applicationId "com.badoo.ribs.samples.helloworld"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
