+++ samples/hello-world/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.samples.helloworld"
+<manifest package="com.badoo.ribs.samples.helloworld"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name="com.samples.helloworld.app.RootActivity">
+        <activity android:name="com.badoo.ribs.samples.helloworld.app.RootActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
