+++ samples/hello-world/src/main/res/layout/activity_root.xml
@@ -4,6 +4,6 @@
     android:id="@+id/root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.samples.helloworld.app.RootActivity">
+    tools:context="com.badoo.ribs.samples.helloworld.app.RootActivity">
 
 </FrameLayout>
